[code] removed code duplicate
